requirements-llamaindex.txt
llama-index-core>=0.12
llama-index-vector-stores-faiss>=0.3
llama-index-retrievers-bm25>=0.5
llama-index-embeddings-ollama>=0.5
llama-index-embeddings-openai>=0.3
